Durak: Show that update() doesn't call paint_event() directly
By using dbgln() we print messages to show, in which order
mousedown_event() and paint_event() are called.
The result is: If update() is called twice in mousedown_event(), the
paint_event is called twice() *after* mousedown_event() returns. That
means, that update() put it's argument the rectangle into a queue, so
that those rectangles can be processed by the paint_event() later.